+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Used user_key in JWT claim from improved security
 - DF-776 Added support for CSV as a data source
 - DF-1075 Added license level in Config - System Info page
+- DF-1106 Added server IP address on system/environment
 
 ### Fixed
 - DF-996 Fixed API Docs to show token refresh endpoints